image-classification/requirements.txt
flask
torch
torchvision
transformers
Pillow
requests
datasets
evaluate
accelerate>=0.26.0
scikit-learn
flask-cors
